Move SecureStoreBackedAutofillStore to autofill-impl module
